Restructure package hierarchy
