+++ folio-spring-i18n/src/main/java/org/folio/spring/i18n/config/TranslationConfiguration.java
@@ -1,4 +1,4 @@
-package org.folio.spring.config;
+package org.folio.spring.i18n.config;
 
 import java.util.Locale;
 import lombok.Getter;
