+++ folio-spring-i18n/src/test/java/org/folio/spring/i18n/model/TranslationMapTest.java
@@ -1,4 +1,4 @@
-package org.folio.spring.model;
+package org.folio.spring.i18n.model;
 
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.is;
